Updated the code so that it can be used also from a notebook. Added an example notebook in the `data` directory.
